Refactor, comments, and bugfix for graphviz writer:
Bug: edges were being added to the Graphviz graph
before the nodes, so the nodes were being
implicitly created prematurely, preventing custom
labels from taking effect and causing the node's
name (which is the uid) to populate the label
instead of the node name.

Refactored code to make things a bit clearer (eg.
function and variable names) and DRY things up.

Added and improved comments to reduce the WTF
factor when trying to re-understand the code.